+++ src/hexen/a_action.c
@@ -200,10 +200,9 @@ void A_PotteryCheck(mobj_t * actor)
     if (!netgame)
     {
         pmo = players[consoleplayer].mo;
-        if (P_CheckSight(actor, pmo) && (abs(R_PointToAngle2(pmo->x,
-                                                             pmo->y, actor->x,
-                                                             actor->y) -
-                                             pmo->angle) <= ANG45))
+        if (P_CheckSight(actor, pmo)
+	  && (abs((int)R_PointToAngle2(pmo->x, pmo->y, actor->x, actor->y)
+           - (int)pmo->angle) <= ANG45))
         {                       // Previous state (pottery bit waiting state)
             P_SetMobjState(actor, actor->state - &states[0] - 1);
         }
@@ -221,11 +220,9 @@ void A_PotteryCheck(mobj_t * actor)
                 continue;
             }
             pmo = players[i].mo;
-            if (P_CheckSight(actor, pmo) && (abs(R_PointToAngle2(pmo->x,
-                                                                 pmo->y,
-                                                                 actor->x,
-                                                                 actor->y) -
-                                                 pmo->angle) <= ANG45))
+            if (P_CheckSight(actor, pmo) 
+              && (abs((int)R_PointToAngle2(pmo->x, pmo->y, actor->x, actor->y)
+               - (int)pmo->angle) <= ANG45))
             {                   // Previous state (pottery bit waiting state)
                 P_SetMobjState(actor, actor->state - &states[0] - 1);
                 return;
